Finished logic for distibution of all functions into diffeent files and deleted functions.py as it's not important anymore
