+++ petrocalc/__init__.py
@@ -1,0 +1,8 @@
+from .units import Units
+from .tree_for_functions import function_tree
+
+__all__ = ["Units", "function_tree"]
+
+"""
+petrocalc - root package that exports function_tree and units
+"""
