+++ petrocalc/pressure_transient_analysis/pta_dimensionless_models.py
@@ -1,0 +1,42 @@
+from __future__ import annotations
+
+import math
+
+__all__ = [
+    "pd_lssihr",
+    "pdw_vwihr",
+    "pdw_vwihrlcpb",
+    "pdw_vwihrlsfb",
+    "pdw_vwihrpcpb",
+    "pdw_vwihrpmb",
+    "pdw_vwihrpsfb"
+]
+
+# PTA dimensionless models
+# FIXME: нет данных для реализации
+def pd_lssihr() -> float:
+    return -1
+
+
+def pdw_vwihr() -> float:
+    return -1
+
+
+def pdw_vwihrlcpb() -> float:
+    return -1
+
+
+def pdw_vwihrlsfb() -> float:
+    return -1
+
+
+def pdw_vwihrpcpb() -> float:
+    return -1
+
+
+def pdw_vwihrpmb() -> float:
+    return -1
+
+
+def pdw_vwihrpsfb() -> float:
+    return -1
