+++ petrocalc/pressure_transient_analysis/pta_models.py
@@ -1,0 +1,39 @@
+from __future__ import annotations
+
+__all__ = [
+    "pw_vwihr",
+    "pw_vwihrlcpb",
+    "pw_vwihrlsfb",
+    "pw_vwihrpcpb",
+    "pw_vwihrpmb",
+    "pw_vwihrpsfb",
+]
+
+
+# Pressure transient analysis
+
+# PTA models
+
+# FIXME: не совпадает с источником
+def pw_vwihr() -> float:
+    return -1
+
+
+def pw_vwihrlcpb() -> float:
+    return -1
+
+
+def pw_vwihrlsfb() -> float:
+    return -1
+
+
+def pw_vwihrpcpb() -> float:
+    return -1
+
+
+def pw_vwihrpmb() -> float:
+    return -1
+
+
+def pw_vwihrpsfb() -> float:
+    return -1
