+++ petrocalc/special_core_analysis/rock_compressibility.py
@@ -1,0 +1,17 @@
+from __future__ import annotations
+
+import math
+
+__all__ = [
+    "cf_newman_l",
+    "cf_newman_s",
+]
+
+# Rock compressibility
+# FIXME: нет коэффициентов для точного вычисения каждой из функций
+def cf_newman_l(phi: float) -> float:
+    return -1
+
+
+def cf_newman_s(phi: float) -> float:
+    return -1
