+++ petrocalc/units.py
@@ -1,0 +1,103 @@
+Units = {
+    "Qi": "bbl/d",
+    "Di": "1/d",
+    "t": "d",
+    "b": "dimensionless",
+    "Dlim": "1/d",
+    "n": "dimensionless",
+    "Qi_intercept": "bbl/d",
+    "Di_intercept": "1/d",
+    "d_inf": "1/d",
+    "time_buildup": "d",
+    "time_plateau": "d",
+    "Q_plateau": "bbl/d",
+    "time_series": "d",
+    "sg_gas": "dimensionless",
+    "api": "°API",
+    "r_s": "scf/STB",
+    "t_f": "°F",
+    "t_r": "°R",
+    "bob": "bbl/STB",
+    "co": "1/psi",
+    "pb": "psia",
+    "p": "psia",
+    'Pb': 'psia',
+    'Uob': 'cP',
+    'rso': 'scf/STB',
+    'rsob': 'scf/STB',
+    'ppr': "dimensionless",
+    'tpr': "dimensionless",
+    'z_factor': "dimensionless",
+    'rswp': 'scf/STB',
+    'rsw': 'scf/STB',
+    'cs': 'g NaCl/L',
+    'salinity': '%wt',
+    'uw1': 'cP',
+    "Ql": "bbl/d",
+    "Ul": "cP",
+    "Rho_l": "lb/ft³",
+    "pipe_id": "in",
+    "pipe_roughness": "dimensionless",
+    "pipe_length": "ft",
+    "pipe_angle": "deg",
+    "P_in": "psia",
+    "P_out": "psia",
+    "Qg": "mscf/d",
+    "Ug": "cP",
+    "T": "°F",
+    "IFTgl": "dyn/cm",
+    "k": "mD",
+    "h": "ft",
+    "bl": "bbl/STB",
+    "ul": "cP",
+    "re": "ft",
+    "rw": "ft",
+    "s": "dimensionless",
+    "porosity": "dimensionless",
+    "ct": "1/psi",
+    "j": "STB/(d·psi)",
+    "pe": "psia",
+    "pwf": "psia",
+    "kz": "mD",
+    "kxy": "mD",
+    "time": "h",
+    "pi": "psia",
+    "h_perf": "ft",
+    "l": "ft",
+    "b_length": "ft",
+    "a": "acre",
+    "rwe": "ft",
+    "xf": "ft",
+    "w": "ft",
+    "k_fracture": "mD",
+    "td": "dimensionless",
+    "rd": "dimensionless",
+    "pd": "dimensionless",
+    "c": "bbl/psi",
+    "cd": "dimensionless",
+    "ld": "dimensionless",
+    "rwd": "dimensionless",
+    "poro": "dimensionless",
+    "mu": "cP",
+    "sw": "dimensionless",
+    "swi": "dimensionless",
+    "sorw": "dimensionless",
+    "krow_swi": "dimensionless",
+    "no": "dimensionless",
+    "krw_sorw": "dimensionless",
+    "nw": "dimensionless",
+    "lo": "dimensionless",
+    "eo": "dimensionless",
+    "to": "dimensionless",
+    "lw": "dimensionless",
+    "ew": "dimensionless",
+    "tw": "dimensionless",
+    "sg": "dimensionless",
+    "x": "dimensionless",
+    "x_values": "dimensionless",
+    "y_values": "dimensionless",
+    "t1": "d",
+    "t2": "d",
+}
+
+__all__ = ["Units"]
