+++ petrocalc/utilities/interpolation.py
@@ -1,0 +1,73 @@
+from __future__ import annotations
+
+__all__ = [
+    "cubic_spline_differentiate",
+    "cubic_spline_integrate",
+    "cubic_spline_integrate_t1_t2",
+    "cubic_spline_interpolate",
+    "cubic_splines_intersection",
+    "data_differentiate",
+    "linear_spline_differentiate",
+    "linear_spline_integrate",
+    "linear_spline_integrate_t1_t2",
+    "linear_spline_interpolate",
+    "linear_splines_intersection",
+    "step_interpolate",
+    "proximal_interpolate",
+]
+
+
+# Utilities
+
+# Interpolation
+# FIXME: не совпадает с источником
+def cubic_spline_differentiate() -> float:
+    return -1
+
+
+def cubic_spline_integrate() -> float:
+    return -1
+
+
+def cubic_spline_integrate_t1_t2() -> float:
+    return -1
+
+
+def cubic_spline_interpolate() -> float:
+    return -1
+
+
+def cubic_splines_intersection() -> float:
+    return -1
+
+
+def data_differentiate() -> float:
+    return -1
+
+
+def linear_spline_differentiate() -> float:
+    return -1
+
+
+def linear_spline_integrate() -> float:
+    return -1
+
+
+def linear_spline_integrate_t1_t2() -> float:
+    return -1
+
+
+def linear_spline_interpolate() -> float:
+    return -1
+
+
+def linear_splines_intersection() -> float:
+    return -1
+
+
+def step_interpolate() -> float:
+    return -1
+
+
+def proximal_interpolate() -> float:
+    return -1
